cut links to godclass, restore original map collector logic to fix test
