+++ core/src/main/java/org/jdbi/v3/core/collector/JdbiCollectors.java
@@ -29,7 +29,11 @@ public class JdbiCollectors implements JdbiConfig<JdbiCollectors> {
     private final List<CollectorFactory> factories = new CopyOnWriteArrayList<>();
 
     public JdbiCollectors() {
-        register(new BuiltInCollectorFactory());
+        // TODO move to BuiltInSupportPlugin
+        register(new MapCollectorFactory());
+        register(new OptionalCollectorFactory());
+        register(new ListCollectorFactory());
+        register(new SetCollectorFactory());
         register(new OptionalPrimitiveCollectorFactory());
         register(new ArrayCollectorFactory());
         register(new EnumSetCollectorFactory());
