+++ core/src/test/java/org/jdbi/v3/core/collector/MapCollectorFactoryTest.java
@@ -14,20 +14,11 @@
 package org.jdbi.v3.core.collector;
 
 import java.lang.reflect.Type;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.LinkedHashMap;
-import java.util.LinkedList;
-import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
 import java.util.SortedMap;
-import java.util.SortedSet;
 import java.util.TreeMap;
-import java.util.TreeSet;
 import java.util.WeakHashMap;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
@@ -42,48 +33,8 @@ import static org.assertj.core.api.Assertions.entry;
 import static org.jdbi.v3.core.generic.GenericTypes.getErasedType;
 import static org.jdbi.v3.core.generic.GenericTypes.resolveMapEntryType;
 
-public class BuiltInCollectorFactoryTest {
-    private BuiltInCollectorFactory factory = new BuiltInCollectorFactory();
-
-    @Test
-    public void optional() {
-        Type optionalString = new GenericType<Optional<String>>() {}.getType();
-        assertThat(factory.accepts(optionalString)).isTrue();
-        assertThat(factory.accepts(Optional.class)).isFalse();
-        assertThat(factory.elementType(optionalString)).contains(String.class);
-
-        Collector<String, ?, Optional<String>> collector = (Collector<String, ?, Optional<String>>) factory.build(optionalString);
-        assertThat(Stream.<String>empty().collect(collector)).isEmpty();
-        assertThat(Stream.of("foo").collect(collector)).contains("foo");
-        assertThatThrownBy(() -> Stream.of("foo", "bar").collect(collector))
-                .isInstanceOf(IllegalStateException.class)
-                .hasMessageContaining("Multiple values");
-    }
-
-    @Test
-    public void collections() {
-        testCollectionType(new GenericType<Collection<String>>(){});
-        testCollectionType(new GenericType<List<String>>(){});
-        testCollectionType(new GenericType<ArrayList<String>>(){});
-        testCollectionType(new GenericType<LinkedList<String>>(){});
-        testCollectionType(new GenericType<Set<String>>(){});
-        testCollectionType(new GenericType<HashSet<String>>(){});
-        testCollectionType(new GenericType<SortedSet<String>>(){});
-        testCollectionType(new GenericType<TreeSet<String>>(){});
-    }
-
-    private <C extends Collection<String>> void testCollectionType(GenericType<C> genericType) {
-        Type containerType = genericType.getType();
-        Class<?> erasedType = getErasedType(containerType);
-        assertThat(factory.accepts(containerType)).isTrue();
-        assertThat(factory.accepts(erasedType)).isFalse();
-        assertThat(factory.elementType(containerType)).contains(String.class);
-
-        Collector<String, ?, C> collector = (Collector<String, ?, C>) factory.build(containerType);
-        assertThat(Stream.of("foo", "bar", "baz").collect(collector))
-                .isInstanceOf(erasedType)
-                .containsOnly("foo", "bar", "baz");
-    }
+public class MapCollectorFactoryTest {
+    private CollectorFactory factory = new MapCollectorFactory();
 
     @Test
     public void maps() {
